+++ src/main/resources/application.properties
@@ -19,4 +19,4 @@ expire.defaultTime=36288000
 expire.products=3600
 
 # tomcat
-server.port=8888
+server.port=9000
